Fix n_split type mismatch between SourceSite and ParticleData
Bug 4: SourceSite::n_split is int64_t but ParticleData::n_split_ was int,
causing implicit narrowing on the round-trip through the shared secondary
bank. Change n_split_ and its accessors to int64_t to match.

+++ include/openmc/particle_data.h
@@ -569,7 +569,7 @@ private:
 
   int64_t n_tracks_ {0}; //!< number of tracks in this particle history
 
-  int n_split_ {0};
+  int64_t n_split_ {0};
   double ww_factor_ {0.0};
 
   int64_t n_progeny_ {0};
@@ -755,8 +755,8 @@ public:
   int& n_event() { return n_event_; }
 
   // Number of times variance reduction has caused a particle split
-  int n_split() const { return n_split_; }
-  int& n_split() { return n_split_; }
+  int64_t n_split() const { return n_split_; }
+  int64_t& n_split() { return n_split_; }
 
   // Particle-specific factor for on-the-fly weight window adjustment
   double ww_factor() const { return ww_factor_; }
